server/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.9.0
Flask-Cors==3.0.8
